[iOS] feat: 텍스트 필드 대신 뷰에 shadow 처리
- 클래스 주석 제거

issue: #18
